Правки схемы объявлений и базовых CRUD операций

+++ backend/app/crud/base.py
@@ -59,8 +59,10 @@ class CRUDBase(Generic[ModelType, CreateSchemaType, UpdateSchemaType]):
         db.refresh(db_obj)
         return db_obj
 
-    def remove(self, db: Session, *, id: int) -> Optional[ModelType]:
+    def remove(self, db: Session, *, id: int) -> ModelType:
         obj = db.query(self.model).get(id)
+        if obj is None:
+            raise ValueError
         db.delete(obj)
         db.commit()
         return obj
